Split up app.ts into middleware files
